fix(change): no longer using Simulate.change
In real applications, your `onChange` handler wont be called if the
value did not change, so the tests shouldn't either. This is another
step toward us not using Simulate at all.

When using `fireEvent.change` make sure that you're setting the value
to change to via: `{target: {value: 'the value'}}` as the second arg.