Add function-based overload to configure()
This commit prepares the way for configure() to be used other than
just for `testIdAttribute`: it exposes the whole config internally
and upgrades the configure() function so that it can take a function
that.

I've introduced explicit tests for this more complex configuration API
and a new documentation section for it. The documentation for the
function-based configure() API could be improved with a real example
once there is a real key for which it would be useful.