feat: expose suggestion data to support tooling
I've added more data to the returned object, to support the development of tooling. Before this change, the returned value would be something like `{ queryName: 'Role' }`, and only by calling the `toString` method, the caller would get more information (the full formatted expression).

With this change, the caller gets more useful data, to build tooling around. Example result:

```
{
  queryName: 'Role',
  queryMethod: 'getByRole',
  queryArgs: ['button', { name: /submit/i }],
  variant: 'get'
}
```

The variant now defaults to `get`, to fix a bug that was causing suggestions like `undefinedByRole(…)` to be generated.