[#610] waitForElementToBeRemoved: fix return type
There's a couple options here, as issue #610 enumerates nicely.

1) Resolve to the type param T

Keeping references to removed elements is a probable source of bugs and
confusing tests. A reference should be easy enough to obtain with other
test utilities; I submit that there's no added value in providing it
here.

2) Resolve to boolean (or literal `true`)

Not a bad choice by any means, but not particularly meaningful (the
fulfillment of the promise is signal enough IMO)

3) Resolve to something empty
This is what I've gone with. We are _literally_ waiting for removal,
after all :)

Fixes #610