feat: Improve performance of findBy*/findAllBy* by not generating informative error messages
in the function passed to waitFor. The work of generating these error
messages almost entirely wasted because all of them, except possibly the
last one (if the test is red), will be discarded by waitFor. Instead,
if the call to waitFor failed, another attempt to find the element(s) is
made, this time generating a useful error message if it also fails.